"Remove UserService dependency from AuthService"
This commit removes the import and instantiation of the UserService from the AuthService file. It also comments out the lines where UserService was being used in the login method. However, the commented out code still references the microservice for users, which suggests that it may still be in use but was accidentally left in this file.