Update package-lock.json, remove plantilla\_rolAdmin.js/json and plantilla\_rolSuperAdmin.json, add new route for creating multiple roles in roleRouter.js, and update roleService.js to include a new function for creating multiple roles with their authorizations
The diff includes changes to package-lock.json, removal of two JavaScript and two JSON files related to role templates, addition of a new route in roleRouter.js for creating multiple roles, and updates to roleService.js to include a new function for creating multiple roles with their authorizations. The new function in roleService.js, createRolesAuthorization, takes in data and an object containing the authorization plantilla and returns an object with a success property set to true and a result property containing the result of the insertOne operation. The new route in roleRouter.js, createRoles, takes in data containing an array of roles to be created and their authorizations, and calls the createRolesAuthorization function in roleService.js to create the roles and their authorizations.